- Added check for rental dates overlap - Added label for determine if a car is on sale - Other minor fixes

+++ src/java/jsf/com/upem/service/mlvUsers/converter/PersonConverter.java
@@ -32,7 +32,7 @@ public class PersonConverter implements Converter {
                 = (RentalController) facesContext.getApplication()
                 .getELResolver().getValue(context.getELContext(), null, "rentalController");
 
-        PersonEntity result = neededBean.retrieveMLVUserByID(Long.parseLong(value));
+        PersonEntity result = neededBean.retrieveMLVUserByID(Integer.parseInt(value));
 
         return result;
     }
